Make it so root workspace layout mode is restored
Previously the layout mode/direction of the root workspace
node was not saved/restored. Now this is done by saving the
whole workspace tree (instead of just the children of the
workspace node) and when the layout is restored the root
workspace node is set correctly and the child nodes are
extracted into a temporary file so that they can be passed
to append_layout on their own.

Closes #33