Ensure cmdline args are always split correctly
Sometimes processes rewrite their own cmdline which often means the args
are no longer separated by the NUL character, which causes issues when
running the command. This commit ensures every element of the cmdline
array is an atomic argument, so that programs that modify their own
cmdline will be restored correctly.