Split single argument commands
Some processes modify their own cmdline so that everything is stuck in
the first argument in shell command form. This is impossible to detect,
but it can be worked around by using shlex.split() on the first argument
if the cmdline has only one argument. The drawback of this is that
valid single arg commands with spaces in the executable path will not
work unless a command mapping is created for them in the config file.